requirements.txt
asn1crypto==0.24.0
cffi==1.11.4
constantly==15.1.0
cryptography==2.1.4
cssselect==1.0.3
flake8==3.5.0
idna==2.6
incremental==17.5.0
Jinja2==2.10
lxml==4.1.1
MarkupSafe==1.0
mock==2.0.0
parsel==1.3.1
pep8==1.7.1
pycparser==2.18
PyDispatcher==2.0.5
pyOpenSSL==17.5.0
python-decouple==3.1
queuelib==1.4.2
requests==2.18.4
Scrapy==1.5.0
six==1.11.0
Twisted==17.9.0
urllib3==1.22
w3lib==1.19.0
zope.interface==4.4.3
